+++ components/checkbox/Group.jsx
@@ -5,6 +5,7 @@ export default React.createClass({
   getDefaultProps() {
     return {
       options: [],
+      defaultValue: [],
       onChange() {},
     };
   },
@@ -17,13 +18,13 @@ export default React.createClass({
   getInitialState() {
     const { value, defaultValue } = this.props;
     return {
-      value: value || defaultValue || [],
+      value: value || defaultValue,
     };
   },
   componentWillReceiveProps(nextProps) {
     if ('value' in nextProps) {
       this.setState({
-        value: nextProps.value,
+        value: nextProps.value || [],
       });
     }
   },
@@ -42,17 +43,19 @@ export default React.createClass({
   },
   render() {
     const options = this.props.options;
-    return <div className="ant-checkbox-group">
-      {
-        options.map(option =>
-          <label className="ant-checkbox-group-item" key={option}>
-            <Checkbox disabled={this.props.disabled}
-            checked={this.state.value.indexOf(option) !== -1}
-            onChange={this.toggleOption.bind(this, option)} />
-            {option}
-          </label>
-        )
-      }
-    </div>;
+    return (
+      <div className="ant-checkbox-group">
+        {
+          options.map(option =>
+            <label className="ant-checkbox-group-item" key={option}>
+              <Checkbox disabled={this.props.disabled}
+                checked={this.state.value.indexOf(option) !== -1}
+                onChange={this.toggleOption.bind(this, option)} />
+              {option}
+            </label>
+          )
+        }
+      </div>
+    );
   },
 });