+++ components/date-picker/PickerMixin.jsx
@@ -13,7 +13,11 @@ export default {
 
   getFormatter() {
     const formats = this.formats = this.formats || {};
-    const format = this.props.format;
+    let format = this.props.format;
+    // Remove time format text when has time-picker in calendar
+    if (this.props.showTime) {
+      format = format.replace('HH:mm:ss', '');
+    }
     if (formats[format]) {
       return formats[format];
     }
@@ -24,16 +28,14 @@ export default {
   parseDateFromValue(value) {
     if (value) {
       if (typeof value === 'string') {
-        return this.getFormatter().parse(value, {locale: this.getLocale()});
+        return this.getFormatter().parse(value, { locale: this.getLocale() });
       } else if (value instanceof Date) {
         let date = new GregorianCalendar(this.getLocale());
         date.setTime(+value);
         return date;
       }
-    } else if (value === null) {
-      return value;
     }
-    return undefined;
+    return value;
   },
 
   // remove input readonly warning