+++ components/dropdown/dropdown.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 import Dropdown from 'rc-dropdown';
 
 export default React.createClass({
-  getDefaultProps: function () {
+  getDefaultProps() {
     return {
       transitionName: 'slide-up',
       prefixCls: 'ant-dropdown',
     };
   },
-  render: function () {
+  render() {
     const { overlay, ...otherProps } = this.props;
     const menu = React.cloneElement(overlay, {
       openTransitionName: 'zoom-big',