+++ components/form/index.jsx
@@ -2,8 +2,20 @@ import Form from './Form';
 import FormItem from './FormItem';
 import ValueMixin from './ValueMixin';
 import Input from '../input';
+import { createForm } from 'rc-form';
 
+Form.create = (o = {}) => {
+  const options = {
+    ...o,
+    fieldNameProp: 'id',
+    fieldMetaProp: '__meta',
+  };
+
+  return createForm(options);
+};
 Form.Item = FormItem;
+
+// @Deprecated
 Form.ValueMixin = ValueMixin;
 
 // 对于 import { Form, Input } from 'antd/lib/form/';