+++ components/radio/group.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import Radio from './radio';
 
 function getCheckedValue(children) {
   let checkedValue = null;
-  React.Children.forEach(children, function (radio) {
+  React.Children.forEach(children, (radio) => {
     if (radio.props && radio.props.checked) {
       checkedValue = radio.props.value;
     }
@@ -12,15 +11,15 @@ function getCheckedValue(children) {
 }
 
 export default React.createClass({
-  getDefaultProps: function () {
+  getDefaultProps() {
     return {
       prefixCls: 'ant-radio-group',
       disabled: false,
-      onChange: function () {
+      onChange() {
       }
     };
   },
-  getInitialState: function () {
+  getInitialState() {
     let props = this.props;
     return {
       value: props.value || props.defaultValue || getCheckedValue(props.children)
@@ -33,30 +32,30 @@ export default React.createClass({
       });
     }
   },
-  render: function () {
-    let props = this.props;
-    let children = React.Children.map(props.children, (radio) => {
-      if (radio.props) {
-        return <Radio
-          key={radio.props.value}
-          {...radio.props}
-          onChange={this.onRadioChange}
-          checked={this.state.value === radio.props.value}
-          disabled={radio.props.disabled || this.props.disabled}
-        />;
-      }
-      return radio;
-    });
-    return (
-      <div className={props.prefixCls}>
-        {children}
-      </div>
-    );
-  },
-  onRadioChange: function (ev) {
+  onRadioChange(ev) {
     this.setState({
       value: ev.target.value
     });
     this.props.onChange(ev);
-  }
+  },
+  render() {
+    const props = this.props;
+    const children = React.Children.map(props.children, (radio) => {
+      if (radio.props) {
+        return React.cloneElement(radio, {
+          key: radio.props.value,
+          ...radio.props,
+          onChange: this.onRadioChange,
+          checked: this.state.value === radio.props.value,
+          disabled: radio.props.disabled || this.props.disabled,
+        });
+      }
+      return radio;
+    });
+    return (
+      <div className={`${props.prefixCls} ${props.prefixCls}-${props.size}`}>
+        {children}
+      </div>
+    );
+  },
 });