+++ components/radio/radio.jsx
@@ -1,5 +1,6 @@
 import Radio from 'rc-radio';
 import React from 'react';
+import classNames from 'classnames';
 
 const AntRadio = React.createClass({
   getDefaultProps() {
@@ -8,17 +9,17 @@ const AntRadio = React.createClass({
     };
   },
   render() {
-    let classString = this.props.className;
-    if (classString) {
-      classString += this.props.checked ? (' ' + classString + '-checked') : '';
-    }
-    if (this.props.disabled) {
-      classString += ' ' + this.props.className + '-disabled';
-    }
+    const { prefixCls, children, checked, disabled, className } = this.props;
+    const classString = classNames({
+      [prefixCls]: true,
+      [prefixCls + '-checked']: checked,
+      [prefixCls + '-disabled']: disabled,
+      [className]: !!className,
+    });
     return (
       <label className={classString}>
         <Radio {...this.props} children={null} />
-        {this.props.children}
+        {children}
       </label>
     );
   }