+++ components/transfer/search.jsx
@@ -12,30 +12,39 @@ class Search extends Component {
     this.props.onChange(e);
   }
 
+  handleClear(e) {
+    e.preventDefault();
+    this.props.handleClear(e);
+  }
+
   render() {
-    const {placeholder, value, prefixCls} = this.props;
-    return <div>
-      <input placeholder={placeholder} className={ prefixCls + ' ant-input' } value={ value } ref="input"
-             onChange={this.handleChange.bind(this)}/>
-      { value && value.length > 0 ?
-        <a href="javascirpt:;" className={ prefixCls + '-action' } onClick={this.props.handleClear}>
-          <Icon type="cross-circle" />
-        </a>
-        : <span className={ prefixCls + '-action' }><Icon type="search" /></span>
-      }
-    </div>;
+    const { placeholder, value, prefixCls } = this.props;
+    return (
+      <div>
+        <input placeholder={placeholder} className={ prefixCls + ' ant-input' } value={ value } ref="input"
+          onChange={this.handleChange.bind(this)}/>
+        { value && value.length > 0 ?
+          <a href="#" className={ prefixCls + '-action' } onClick={this.handleClear.bind(this)}>
+            <Icon type="cross-circle" />
+          </a>
+          : <span className={ prefixCls + '-action' }><Icon type="search" /></span>
+        }
+      </div>
+    );
   }
 }
 
 Search.defaultProps = {
   placeholder: '请输入搜索内容',
   onChange: noop,
+  handleClear: noop,
 };
 
 Search.propTypes = {
   prefixCls: PropTypes.string,
   placeholder: PropTypes.string,
-  onChange: PropTypes.func
+  onChange: PropTypes.func,
+  handleClear: PropTypes.func,
 };
 
 export default Search;