+++ webpack.config.js
@@ -4,15 +4,18 @@ var path = require('path');
 var pkg = require('./package');
 
 var entry = {};
-entry['index'] = './scripts/importCss.js';
 entry['demo'] = './scripts/demo.js';
 
 module.exports = {
   entry: entry,
 
+  cache: true,
+
   resolve: {
     extensions: ['', '.js', '.jsx']
   },
+
+  noParse: /_site|node_modules/,
 
   output: {
     path: path.join(process.cwd(), 'dist'),
@@ -34,16 +37,19 @@ module.exports = {
       }
     }, {
       test: /\.json$/,
+      exclude: /node_modules/,
       loader: 'json-loader'
     }, {
       test: /\.less$/,
+      exclude: /node_modules/,
       loader: ExtractTextPlugin.extract(
-        'css?sourceMap&-minimize!' + 'autoprefixer-loader!' + 'less?sourceMap'
+        'css?sourceMap&-minimize!' + 'postcss-loader!' + 'less?sourceMap'
       )
     }, {
       test: /\.css$/,
+      exclude: /node_modules/,
       loader: ExtractTextPlugin.extract(
-        'css?sourceMap&-minimize!' + 'autoprefixer-loader'
+        'css?sourceMap&-minimize!' + 'postcss-loader'
       )
     }]
   },